Repaired (broken in previous commits) possibility to build without OpenGL
